domparsing/domparser.html

<!DOCTYPE html>
<html>
 <head>
  <title>DOM Parser</title>
  <script src="../resources/testharness.js"></script>
  <script src="../resources/testharnessreport.js"></script>
 </head>
 <body>
  <script>
  var emptyHTML = "<!DOCTYPE html><html xmlns=\"http://www.w3.org/1999/xhtml\"><head><title>empty</title></head><body></body></html>";
  var soup = "<!DOCTYPE foo></><foo></multiple></>";
  var emptyXML = "<!DOCTYPE foo><foo xmlns=\"test_ns\" />";
  test(function() {
    var htmldoc = new DOMParser().parseFromString(soup, "text/html");
    assert_equals(htmldoc.documentElement.namespaceURI, "http://www.w3.org/1999/xhtml");
  }, "DOMParser parses HTML tag soup with no problems");
  test(function() {
    assert_throws(new SyntaxError(), function() { new DOMParser().parseFromString(soup, "application/xml"); });
  }, "DOMParser throws an exception when parsing tag soup as XML.");
  test(function() {
    var htmldocDOMParsed = new DOMParser().parseFromString(emptyHTML, "text/html");
    var xhtmldocDOMParsed = new DOMParser().parseFromString(emptyHTML, "application/xhtml+xml" );
    
    assert_equals(htmldocDOMParsed.location, xhtmldocDOMParsed.location);
    assert_equals(htmldocDOMParsed.URL, xhtmldocDOMParsed.URL);
    assert_equals(htmldocDOMParsed.documentURI, xhtmldocDOMParsed.documentURI);
  }, "Documents parsed as HTML or XML have the same origin properties regardless");
  // Test what happens when unknown string is provided!
  test(function() {
    var htmldocDOMParsed = new DOMParser().parseFromString(emptyHTML, "text/html");
    assert_equals(htmldocDOMParsed.URL, document.URL);
  }, "DOMParser parsed documents have a URL property equal to the document's URL");
  test(function() {
    var htmldocDOMParsed = new DOMParser().parseFromString(emptyHTML, "text/html");
    assert_equals(htmldocDOMParsed.location, null);
  }, "DOMParser parsed documents have a location equal to null [HTML5]");
  test(function() {
    var htmldocDOMParsed = new DOMParser().parseFromString(emptyHTML, "text/html");
    assert_equals(htmldocDOMParsed.documentURI, document.URL);
  }, "DOMParser parsed documents have a documentURL equal to the document's URL [DOM4]");
  
  </script>
  <div id="log"></div>
 </body>
</html>
